Prevent XSOM Parser from processing the same schema multiple times and fail
Fix for [JAXB-875](https://github.com/javaee/jaxb-v2/issues/875):

  * re-implement resolution of schema imports and includes with detailed test case
  * added a logger to the NGCCRuntime engine, its not 1995 anymore
  * fixed javadoc param in NGCCRuntime
  * added mock to pom
  * added schemas folder as test sources

This change prevents XSOM from loading same schema again if it is imported multiple times in other schemas. If you get the dreaded `org.xml.sax.SAXParseException: 'YourType' is already defined` error when compiling your set of schemas, then this fix is for you. The runtime now also supports the import of schema using the namespace attribute only. The parser will treat the value of the namespace attribute as publicId and will delegate to the entityResolver for resolution. Provide a properly configured entity resolver to the XSOM parser and things should start working.